requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
idna==3.7
Pygments==2.18.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
